fresh-and-go/services/iot/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-dateutil==2.8.2
jsonschema==4.20.0
pydantic==2.5.0
